# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateReviewRequest:
    properties:
      comment:
        maxLength: 1000
        minLength: 1
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      reviewer_name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - comment
    - rating
    - reviewer_name
    type: object
...
  dto.TopUpRequest:
    properties:
      amount:
        type: integer
    required:
    - amount
    type: object
...
  dto.UpsertStoreRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 150
        minLength: 3
        type: string
    required:
    - name
    type: object
  response.ErrorCode:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - UNPROCESSABLE_ENTITY
    - INTERNAL_SERVER_ERROR
    - VALIDATION_ERROR
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    type: string
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeNotFound
    - ErrCodeConflict
    - ErrCodeUnprocessable
    - ErrCodeInternalServerError
    - ErrCodeValidation
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      error_code:
        $ref: '#/definitions/response.ErrorCode'
      list_data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      status:
        type: boolean
    type: object
info:
  contact:
    name: SEAPEDIA Dev
  description: |-
    Backend API marketplace multi-role (Admin, Seller, Buyer, Driver) — COMPFEST 18.
    Versi yang tampil di UI diambil dari config.Version (di-set saat runtime).
    Login pakai email. Otorisasi mengikuti ACTIVE ROLE pada JWT (pilih via /auth/select-role).
  termsOfService: http://swagger.io/terms/
  title: SEAPEDIA API
  version: "1.0"
paths:
...
  /buyer/wallet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Saldo wallet (buyer)
      tags:
      - Buyer
  /buyer/wallet/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Jumlah top-up
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Top-up dummy (buyer)
      tags:
      - Buyer
  /buyer/wallet/transactions:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Riwayat transaksi wallet (buyer)
      tags:
      - Buyer
...
  /reviews:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Daftar review aplikasi (publik)
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Review
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Kirim review aplikasi (publik, guest boleh)
      tags:
      - Reviews
...
  /seller/store:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail toko sendiri (seller)
      tags:
      - Seller
    post:
      consumes:
      - application/json
      parameters:
      - description: Data toko
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat/ubah toko sendiri (seller)
      tags:
      - Seller
  /stores:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Daftar toko (publik)
      tags:
      - Public
  /stores/{id}:
    get:
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail toko (publik)
      tags:
      - Public
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"